refactor: relocate tools/observe/ → src/Core.TypeScript/observe/
Move the entire observe loop (34 TS files + data files) from tools/
into src/Core.TypeScript/ where it belongs alongside its siblings
(ferry-throttler, agent-bus, workflow-engine, zeta-id).

Import path fixes:
- Within observe: ../../src/Core.TypeScript/X → ../X (now siblings)
- From observe to tools/: ../accelerator, ../backlog → ../../../tools/
- From agent-bus/clean.ts: ../../../tools/observe → ../observe
- From poll-pr-gate-batch.ts: ../observe → ../../src/Core.TypeScript/observe

Runtime path fixes for data files that use relative filesystem paths
(context-cost-trend.jsonl, token-calibration.json, etc).

178 tests pass, zero tsc diagnostics.
