+++ src/Core.TypeScript/observe/context-cost-trend.jsonl
