+++ src/Core.TypeScript/observe/event-sink-folder.test.ts
@@ -11,8 +11,8 @@ import { afterEach, beforeEach, describe, expect, it } from "bun:test";
 import { existsSync, mkdtempSync, readFileSync, rmSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
-import { unpack } from "../../src/Core.TypeScript/zeta-id/zeta-id";
-import { Category, type ZetaId } from "../../src/Core.TypeScript/zeta-id/types";
+import { unpack } from "../zeta-id/zeta-id";
+import { Category, type ZetaId } from "../zeta-id/types";
 import { execute } from "./execute";
 import { coauthorFor, folderSink, gitCommitToMain, mintObserveEventIdHex, type CommitOutcome, type EventEnvelope } from "./event-sink-folder";
 import type { NextAction, World } from "./observe";
