+++ src/Core.TypeScript/observe/golden-vectors.json
