+++ src/Core.TypeScript/observe/golden-vectors.ts
