+++ src/Core.TypeScript/observe/grammar-16.test.ts
