+++ src/Core.TypeScript/observe/observe.ts
@@ -59,8 +59,8 @@
  * later, the work-hours KPI overlay (Max — expectations, not a time-lock).
  */
 
-import { chooseIndex, ollamaBackend, type ModelBackend } from "../accelerator/local-llm";
-import type { FourCornerOwnership } from "../../src/Core.TypeScript/workflow-engine/types";
+import { chooseIndex, ollamaBackend, type ModelBackend } from "../../../tools/accelerator/local-llm";
+import type { FourCornerOwnership } from "../workflow-engine/types";
 
 /** One backlog item, classified to just what the controller needs to decide. */
 export interface BacklogItem {
