+++ src/Core.TypeScript/observe/pointer-check.ts
